refactor: move resolveEarlyReportID to IOUUtils
