+++ tests/unit/useResetIOUTypeReportIDTest.ts
@@ -1,4 +1,4 @@
-import {resolveEarlyReportID} from '@hooks/useResetIOUType';
+import {resolveEarlyReportID} from '@libs/IOUUtils';
 import CONST from '@src/CONST';
 import type {Participant} from '@src/types/onyx/IOU';
